+++ record/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 6.1.0
 * feat: Add AudioInterruptionMode to `RecordConfig`.
 * feat: Add stream buffer size option.
+* feat: Allow background recording on Android & iOS.
 * chore: Updated transitive dependencies.
 * chore: Update minimum supported SDK version to Flutter 3.24.0/Dart 3.5.0.