+++ record_ios/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0
 * feat: Add AudioInterruptionMode to `RecordConfig`.
 * feat: Add stream buffer size option.
+* feat: Allow background recording.
 
 ## 1.0.0
 * chore: Support Swift Package Manager.